Fix imports and move lib folder to src for proper module resolution
