+++ src/lib/ai/portfolio-generator.ts
